change colors + move files in their directories
